refactor: jwt 생성 및 검증기 이름 변경 & 위치 이동
